+++ app/assets/scripts/app.js
@@ -409,7 +409,7 @@ function( angular , ngAnimate , minireel     , account     , login , portal , c6
                 };
 
                 this.find = function(type, id) {
-                    return $http.get(url('card/' + id))
+                    return $http.get(url('card/' + id), { cache: true })
                         .then(pick('data'))
                         .then(convertCardForEditor)
                         .then(putInArray);
@@ -518,7 +518,7 @@ function( angular , ngAnimate , minireel     , account     , login , portal , c6
                 };
 
                 this.find = function(type, id) {
-                    return $http.get(url('customer/' + id))
+                    return $http.get(url('customer/' + id), { cache: true })
                         .then(pick('data'))
                         .then(this.decorate)
                         .then(putInArray);
@@ -551,7 +551,7 @@ function( angular , ngAnimate , minireel     , account     , login , portal , c6
                 };
 
                 this.find = function(type, id) {
-                    return $http.get(url('advertiser/' + id))
+                    return $http.get(url('advertiser/' + id), { cache: true })
                         .then(pick('data'))
                         .then(putInArray);
                 };
@@ -707,7 +707,7 @@ function( angular , ngAnimate , minireel     , account     , login , portal , c6
                 };
 
                 this.find = function(type, id) {
-                    return $http.get(url('campaign/' + id))
+                    return $http.get(url('campaign/' + id), { cache: true })
                         .then(pick('data'))
                         .then(this.decorateCampaign)
                         .then(putInArray);